requirements.txt
openai==0.27.4
pinecone-client==2.2.1
python-dotenv==1.0.0
tiktoken==0.3.3
termcolor==2.2.0
discord.py==2.3.0
psycopg==3.1.9
pgvector==0.1.8
